Bug#22848937 FIX MORE ERRORS REPORTED BY UBSAN - GOOGLE PROTOBUF
protobuf-2.6.0/src/google/protobuf/io/coded_stream.cc:634
null pointer passed as argument 1, which is declared to never be null

protobuf-2.6.0/src/google/protobuf/io/printer.cc:181
null pointer passed as argument 1, which is declared to never be null

protobuf-2.6.0/src/google/protobuf/stubs/strutil.cc:908
signed integer overflow: 42 * -100000000 cannot be represented in type 'int'

protobuf-2.6.0/src/google/protobuf/text_format.cc:1113
null pointer passed as argument 1, which is declared to never be null

protobuf-2.6.0/src/google/protobuf/wire_format_lite.h:650
left shift of negative value -128

+++ extra/protobuf/protobuf-2.6.0/src/google/protobuf/wire_format_lite.h
@@ -647,7 +647,7 @@ inline int32 WireFormatLite::ZigZagDecode32(uint32 n) {
 
 inline uint64 WireFormatLite::ZigZagEncode64(int64 n) {
   // Note:  the right-shift must be arithmetic
-  return (n << 1) ^ (n >> 63);
+  return (static_cast<uint64>(n) << 1) ^ (n >> 63);
 }
 
 inline int64 WireFormatLite::ZigZagDecode64(uint64 n) {
